allLessonsReducer, addLessonReducer, editLessonReducer were added
