+++ src/models/lessons/actions.js
@@ -8,6 +8,21 @@ export const fetchLessons = () => ({
   type: actionsTypes.FETCH_LESSONS,
 });
 
+export const addLesson = (data) => ({
+  type: actionsTypes.ADD_LESSON,
+  payload: {
+    data,
+  },
+});
+
+export const editLesson = (data, id) => ({
+  type: actionsTypes.EDIT_LESSON,
+  payload: {
+    data,
+    id,
+  },
+});
+
 function* fetchLessonsAsync() {
   try {
     yield put({ type: actionsTypes.LOADING_LESSONS_STARTED });
@@ -29,17 +44,6 @@ function* fetchLessonsAsync() {
     });
   }
 }
-
-export function* fetchLessonsWatcher() {
-  yield takeLatest(actionsTypes.FETCH_LESSONS, fetchLessonsAsync);
-}
-
-export const addLesson = (data) => ({
-  type: actionsTypes.ADD_LESSON,
-  payload: {
-    data,
-  },
-});
 
 function* addLessonAsync(lessonData) {
   try {
@@ -63,18 +67,6 @@ function* addLessonAsync(lessonData) {
   }
 }
 
-export function* addLessonWatcher() {
-  yield takeEvery(actionsTypes.ADD_LESSON, addLessonAsync);
-}
-
-export const editLesson = (data, id) => ({
-  type: actionsTypes.EDIT_LESSON,
-  payload: {
-    data,
-    id,
-  },
-});
-
 function* editLessonAsync(editData) {
   try {
     yield put({ type: actionsTypes.EDITING_STARTED });
@@ -95,6 +87,14 @@ function* editLessonAsync(editData) {
       },
     });
   }
+}
+
+export function* fetchLessonsWatcher() {
+  yield takeLatest(actionsTypes.FETCH_LESSONS, fetchLessonsAsync);
+}
+
+export function* addLessonWatcher() {
+  yield takeEvery(actionsTypes.ADD_LESSON, addLessonAsync);
 }
 
 export function* editLessonWatcher() {
