+++ src/models/lessons/index.js
@@ -2,4 +2,4 @@ export { lessonsReducer } from './reducer.js';
 export {
   fetchLessons, editLesson, addLesson, lessonsWatcher,
 } from './actions.js';
-export { lessonsSelector } from './selectors.js';
+export { lessonsSelector, addLessonSelector, editLessonSelector } from './selectors.js';
