+++ src/models/lessons/reducers/all-lessons-reducer.js
@@ -1,0 +1,36 @@
+import * as actionsTypes from '../action-types.js';
+
+const INITIAL_STATE = {
+  data: [],
+  isLoading: false,
+  isLoaded: false,
+  error: '',
+};
+
+export const allLessonsReducer = (state = INITIAL_STATE, action) => {
+  switch (action.type) {
+    case actionsTypes.LOADING_LESSONS_STARTED:
+      return {
+        ...state,
+        isLoading: true,
+        error: '',
+      };
+    case actionsTypes.LOADING_LESSONS_FINISHED:
+      return {
+        ...state,
+        isLoading: false,
+        loaded: true,
+        data: action.payload.data,
+        error: '',
+      };
+    case actionsTypes.LOADING_LESSONS_FAILED:
+      return {
+        ...state,
+        isLoading: false,
+        loaded: false,
+        error: action.payload.error,
+      };
+    default:
+      return state;
+  }
+};
