+++ src/models/lessons/selectors.js
@@ -1,1 +1,3 @@
-export const lessonsSelector = (state) => state.models.lessons;
+export const lessonsSelector = (state) => state.models.lessons.lessons;
+export const addLessonSelector = (state) => state.models.lessons.addLesson;
+export const editLessonSelector = (state) => state.models.lessons.editLesson;
